+++ core/src/test/java/cucumber/table/FromDataTableTest.java
@@ -1,13 +1,18 @@
-package cucumber.runtime.java;
+package cucumber.table;
 
 import com.thoughtworks.xstream.annotations.XStreamConverter;
+import com.thoughtworks.xstream.converters.Converter;
+import com.thoughtworks.xstream.converters.MarshallingContext;
+import com.thoughtworks.xstream.converters.UnmarshallingContext;
 import com.thoughtworks.xstream.converters.javabean.JavaBeanConverter;
-import cucumber.annotation.DateFormat;
+import com.thoughtworks.xstream.io.HierarchicalStreamReader;
+import com.thoughtworks.xstream.io.HierarchicalStreamWriter;
+import cucumber.DateFormat;
 import cucumber.runtime.CucumberException;
+import cucumber.runtime.ParameterType;
 import cucumber.runtime.StepDefinition;
 import cucumber.runtime.StepDefinitionMatch;
 import cucumber.runtime.converters.LocalizedXStreams;
-import cucumber.table.DataTable;
 import gherkin.I18n;
 import gherkin.formatter.Argument;
 import gherkin.formatter.model.Comment;
@@ -24,13 +29,12 @@ import java.util.Date;
 import java.util.List;
 import java.util.Map;
 import java.util.TimeZone;
-import java.util.regex.Pattern;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
 import static junit.framework.Assert.assertEquals;
 
-public class JavaTableConversionTest {
+public class FromDataTableTest {
     @Rule
     public ExpectedException thrown = ExpectedException.none();
 
@@ -40,6 +44,7 @@ public class JavaTableConversionTest {
     public static class StepDefs {
         public List<UserPojo> listOfPojos;
         public List<UserBean> listOfBeans;
+        public List<UserWithNameField> listOfUsersWithNameField;
         public List<List<Double>> listOfListOfDoubles;
         public List<Map<String, String>> listOfMapsOfStringToString;
         public List<Map<String, Object>> listOfMapsOfStringToObject;
@@ -52,6 +57,10 @@ public class JavaTableConversionTest {
 
         public void listOfBeans(@DateFormat("yyyy-MM-dd") List<UserBean> listOfBeans) {
             this.listOfBeans = listOfBeans;
+        }
+
+        public void listOfUsersWithNameField(@DateFormat("yyyy-MM-dd") List<UserWithNameField> listOfUsersWithNameField) {
+            this.listOfUsersWithNameField = listOfUsersWithNameField;
         }
 
         public void listOfListOfDoubles(List<List<Double>> listOfListOfDoubles) {
@@ -92,6 +101,15 @@ public class JavaTableConversionTest {
         Method m = StepDefs.class.getMethod("listOfBeans", List.class);
         StepDefs stepDefs = runStepDef(m, listOfDatesWithHeader());
         assertEquals(sidsBirthday(), stepDefs.listOfBeans.get(0).getBirthDate());
+    }
+
+    @Test
+    public void converts_table_to_list_of_class_with_special_fields() throws Throwable {
+        Method m = StepDefs.class.getMethod("listOfUsersWithNameField", List.class);
+        StepDefs stepDefs = runStepDef(m, listOfDatesAndNamesWithHeader());
+        assertEquals(sidsBirthday(), stepDefs.listOfUsersWithNameField.get(0).birthDate);
+        assertEquals("Sid", stepDefs.listOfUsersWithNameField.get(0).name.first);
+        assertEquals("Vicious", stepDefs.listOfUsersWithNameField.get(0).name.last);
     }
 
     @Test
@@ -152,7 +170,7 @@ public class JavaTableConversionTest {
 
     private StepDefs runStepDef(Method method, List<DataTableRow> rows) throws Throwable {
         StepDefs stepDefs = new StepDefs();
-        StepDefinition stepDefinition = new JavaStepDefinition(method, Pattern.compile("whatever"), new SingletonFactory(stepDefs));
+        StepDefinition stepDefinition = new DirectStepDef(stepDefs, method);
 
         Step stepWithRows = new Step(NO_COMMENTS, "Given ", "something", 10, rows, null);
 
@@ -166,6 +184,13 @@ public class JavaTableConversionTest {
         List<DataTableRow> rows = new ArrayList<DataTableRow>();
         rows.add(new DataTableRow(NO_COMMENTS, asList("Birth Date"), 1));
         rows.add(new DataTableRow(NO_COMMENTS, asList("1957-05-10"), 2));
+        return rows;
+    }
+
+    private List<DataTableRow> listOfDatesAndNamesWithHeader() {
+        List<DataTableRow> rows = new ArrayList<DataTableRow>();
+        rows.add(new DataTableRow(NO_COMMENTS, asList("Birth Date", "Name"), 1));
+        rows.add(new DataTableRow(NO_COMMENTS, asList("1957-05-10", "Sid Vicious"), 2));
         return rows;
     }
 
@@ -199,6 +224,78 @@ public class JavaTableConversionTest {
 
         public void setBirthDate(Date birthDate) {
             this.birthDateX = birthDate;
+        }
+    }
+
+    public static class UserWithNameField {
+        public Name name;
+        public Date birthDate;
+    }
+
+    @XStreamConverter(NameConverter.class)
+    public static class Name {
+        public String first;
+        public String last;
+    }
+
+    public static class NameConverter implements Converter {
+        @Override
+        public void marshal(Object source, HierarchicalStreamWriter writer, MarshallingContext context) {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public Object unmarshal(HierarchicalStreamReader reader, UnmarshallingContext context) {
+            Name name = new Name();
+            String[] firstLast = reader.getValue().split(" ");
+            name.first = firstLast[0];
+            name.last = firstLast[1];
+            return name;
+        }
+
+        @Override
+        public boolean canConvert(Class type) {
+            return type.equals(Name.class);
+        }
+    }
+
+    private class DirectStepDef implements StepDefinition {
+        private final Object target;
+        private final Method method;
+
+        public DirectStepDef(Object target, Method method) {
+            this.target = target;
+            this.method = method;
+        }
+
+        @Override
+        public List<Argument> matchedArguments(Step step) {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public String getLocation() {
+            return getClass().getName();
+        }
+
+        @Override
+        public List<ParameterType> getParameterTypes() {
+            return ParameterType.fromMethod(method);
+        }
+
+        @Override
+        public void execute(I18n i18n, Object[] args) throws Throwable {
+            method.invoke(target, args);
+        }
+
+        @Override
+        public boolean isDefinedAt(StackTraceElement stackTraceElement) {
+            return false;
+        }
+
+        @Override
+        public String getPattern() {
+            throw new UnsupportedOperationException();
         }
     }
 }